Added progress bar for home page and alert dialog for delete account

+++ app/src/main/java/com/podcastlist/ui/screen/edit_account/EditAccountScreen.kt
@@ -1,10 +1,14 @@
 package com.podcastlist.ui.screen.edit_account
 
-import androidx.compose.material.ButtonDefaults
-import androidx.compose.material.MaterialTheme
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material.*
+import androidx.compose.runtime.*
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.podcastlist.R
 import com.podcastlist.auth.AuthButton
@@ -22,6 +26,7 @@ fun EditAccountScreen(
 ) {
     val uiState by viewModel.uiState
     viewModel.snackbarManager = snackbarManager
+    var openDialog by remember { mutableStateOf(false) }
 
     FormColumn {
         EmailField(value = uiState.email, onEmailChange = viewModel::onEmailChange)
@@ -56,7 +61,44 @@ fun EditAccountScreen(
                 contentColor = MaterialTheme.colors.onError
             )
         ) {
-            viewModel.deleteAccount(navigateToHome)
+            openDialog = true
+        }
+
+        if (openDialog) {
+            AlertDialog(
+                onDismissRequest = { openDialog = false },
+                title = { Text("Are you sure you want to delete the account?") },
+                text = {
+                    Text("Beware that neither the account nor any its associated data can be recovered after deleting the account.")
+                },
+                buttons = {
+                    Row (
+                        modifier = Modifier.padding(bottom = 10.dp, start = 20.dp, end = 20.dp),
+                        horizontalArrangement = Arrangement.Center
+                    ) {
+                        Button(
+                            onClick = {
+                                openDialog = false
+                            }
+                        ) {
+                            Text("Cancel")
+                        }
+                        Spacer(modifier = Modifier.weight(1f))
+                        Button(
+                            onClick = {
+                                openDialog = false
+                                viewModel.deleteAccount(navigateToHome)
+                            },
+                            colors = ButtonDefaults.buttonColors(
+                                backgroundColor = MaterialTheme.colors.error,
+                                contentColor = MaterialTheme.colors.onError
+                            )
+                        ) {
+                            Text("Yes, delete it")
+                        }
+                    }
+                }
+            )
         }
     }
 }
